[refactor] 패키지 구조 정리 및 key.yml 분리
