+++ src/main/java/com/ssafy/roscamback/config/RedisConfig.java
@@ -1,7 +1,6 @@
 package com.ssafy.roscamback.config;
 
 
-import com.ssafy.roscamback.dto.MessageDto;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,11 +9,9 @@ import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
-@Profile("!test")
 public class RedisConfig {
 
     @Value("${spring.data.redis.host}")
