--- src/main/java/com/ssafy/roscamback/dto/MessageDto.java
@@ -1,8 +1,0 @@
-package com.ssafy.roscamback.dto;
-
-public class MessageDto {
-
-    private String userId;
-    private String otherId;
-    private String data;
-}
