+++ src/main/java/com/ssafy/roscamback/service/ServiceImpl.java
@@ -3,9 +3,9 @@ package com.ssafy.roscamback.service;
 import com.ssafy.roscamback.dto.response.ChatAnalysisResponse;
 import com.ssafy.roscamback.module.ChatSaveEvent;
 import com.ssafy.roscamback.dto.request.AnalyzeRequest;
-import com.ssafy.roscamback.dto.request.ThreatEntry;
+import com.ssafy.roscamback.dto.request.url.ThreatEntry;
 import com.ssafy.roscamback.dto.response.AnalyzeResponse;
-import com.ssafy.roscamback.dto.response.CheckResponse;
+import com.ssafy.roscamback.dto.response.url.CheckResponse;
 import com.ssafy.roscamback.util.ApiChecker;
 import com.ssafy.roscamback.util.FastApiUtil;
 import java.util.concurrent.CompletableFuture;
