+++ src/main/resources/application.yml
@@ -13,11 +13,6 @@ spring:
   mvc:
     async:
       request-timeout: 600000 #600초
-  data:
-    redis:
-      port: 6379
-      host: localhost
-      password:
 
   jpa:
     hibernate:
@@ -29,8 +24,3 @@ spring:
   sql:
     init:
       mode: always
-  datasource:
-    url: jdbc:mysql://localhost/test?serverTimezone=Asia/Seoul
-    username: root
-    password: 1q2w3e4r
-    driver-class-name: com.mysql.cj.jdbc.Driver
